pyproject.toml
[project]
name = "price-monitoring"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cachetools>=5.5.2",
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.0",
]
